+++ src/sessionstorer/test/initialize.php
@@ -6,7 +6,7 @@ define("SESSION_STORER_COOKIE_NAME_SESSION","%session_name%");
 define("SESSION_STORER_COOKIE_NAME_CHECK","check");
 define("SESSION_STORER_SESSION_MAX_LIFETIME",60 * 60 * 24 * 1);  // a day
 define("CURRENT_TIME",time());
-require("../../stringbuffer/stringbuffer.php");
+require("../../stringbuffer/load.php");
 require("../../http/load.php");
 require("../../string4/load.php");
 require("../sessionstorer.php");